# api/swagger.yaml
openapi: 3.0.3
info:
  title: Viral Content Scraper API
  description: |
    # 🚀 API de Scraping Inteligente de Conteúdo Viral
    
    Uma API completa para coleta, análise e geração de insights sobre conteúdo viral em múltiplas plataformas sociais.
    
    ## 🎯 Funcionalidades Principais
    
    - **Scraping Inteligente**: Coleta automatizada de conteúdo viral
    - **Análise com IA**: Processamento avançado com GPT-4 Vision
    - **Templates Virais**: Extração e adaptação de padrões de sucesso
    - **Análise de Perfis**: Análise completa de perfis Instagram
    - **Tendências**: Identificação de padrões virais em tempo real
    - **Administração**: Gerenciamento completo do sistema
    - **Webhooks**: Notificações em tempo real
    
    ## 🔐 Autenticação
    
    A API utiliza autenticação JWT Bearer Token:
    ```
    Authorization: Bearer <seu_token_jwt>
    ```
    
    ## 📊 Rate Limiting
    
    - **Usuário Padrão**: 1000 requests/hora, 100 requests/minuto
    - **Usuário Premium**: 5000 requests/hora, 500 requests/minuto
    - **Usuário Business**: 15000 requests/hora, 1500 requests/minuto
    
    ## 🌐 Base URL
    
    - **Produção**: `https://api.viralscraper.com/api/v1`
    - **Desenvolvimento**: `http://localhost:5000/api/v1`
    
  version: 1.0.0
  contact:
    name: Manus AI
    email: support@viralscraper.com
    url: https://viralscraper.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
...
security:
  - BearerAuth: []
...
paths:
...
  /analysis/content/{content_id}:
    post:
      tags: [Analysis]
      summary: Analisar conteúdo específico
      description: Realiza análise completa de um conteúdo (sentiment, visual, métricas)
      parameters:
        - name: content_id
          in: path
          required: true
          schema:
            type: string
          description: ID único do conteúdo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                analysis_types:
                  type: array
                  items:
                    type: string
                    enum: [sentiment, visual, metrics, all]
                  default: [all]
                include_recommendations:
                  type: boolean
                  default: true
                cache_results:
                  type: boolean
                  default: true
      responses:
        '200':
          description: Análise concluída com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /trends/viral:
    get:
      tags: [Trends]
      summary: Conteúdo viral em tempo real
      description: Detecta conteúdo viral com base em critérios avançados
      parameters:
        - name: platform
          in: query
          schema:
            type: string
            enum: [instagram, tiktok, youtube, all]
        - name: period
          in: query
          schema:
            type: string
            enum: [1h, 6h, 24h, 7d]
            default: 24h
        - name: content_type
          in: query
          schema:
            type: string
            enum: [reel, post, carousel, video]
        - name: min_viral_score
          in: query
          schema:
            type: integer
            minimum: 0
            maximum: 100
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Conteúdo viral encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ViralTrendsResponse'
...
  /templates:
    get:
      tags: [Templates]
      summary: Listar templates
      description: Busca templates com filtros avançados
      parameters:
        - name: content_type
          in: query
          schema:
            type: string
            enum: [carousel, reel, post, story]
        - name: min_viral_score
          in: query
          schema:
            type: integer
            minimum: 0
            maximum: 100
        - name: platform
          in: query
          schema:
            type: string
            enum: [instagram, tiktok, linkedin, youtube]
        - name: search_tags
          in: query
          schema:
            type: string
            description: Tags separadas por vírgula
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [viral_score, created_at, template_quality_score]
            default: viral_score
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [desc, asc]
            default: desc
      responses:
        '200':
          description: Lista de templates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplatesListResponse'

    post:
      tags: [Templates]
      summary: Criar template personalizado
      description: Cria um novo template baseado em especificações
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTemplateRequest'
      responses:
        '201':
          description: Template criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateResponse'
  /templates/{template_id}:
    get:
      tags: [Templates]
      summary: Buscar template específico
      parameters:
        - name: template_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detalhes do template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateResponse'

    delete:
      tags: [Templates]
      summary: Deletar template
      parameters:
        - name: template_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Template deletado com sucesso
...
  /profiles/{username}:
    get:
      tags: [Profiles]
      summary: Buscar análise de perfil
      description: Recupera análise mais recente de um perfil
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
        - name: platform
          in: query
          schema:
            type: string
            enum: [instagram, tiktok, youtube, linkedin]
            default: instagram
      responses:
        '200':
          description: Análise do perfil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileAnalysisResponse'
  /profiles:
    get:
      tags: [Profiles]
      summary: Listar análises de perfis
      description: Lista todas as análises de perfis com filtros
      parameters:
        - name: platform
          in: query
          schema:
            type: string
            enum: [instagram, tiktok, youtube, linkedin]
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: sort_by
          in: query
          schema:
            type: string
            enum: [analyzed_at, followers, viral_rate]
            default: analyzed_at
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [desc, asc]
            default: desc
      responses:
        '200':
          description: Lista de análises de perfis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfilesListResponse'
...
  /webhooks:
    get:
      tags: [Webhooks]
      summary: Listar webhooks
      description: Lista todos os webhooks registrados
      responses:
        '200':
          description: Lista de webhooks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhooksListResponse'

    post:
      tags: [Webhooks]
      summary: Registrar webhook
      description: Registra novo webhook para receber notificações
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  format: uri
                  description: URL que receberá as notificações
                events:
                  type: array
                  items:
                    type: string
                    enum: [content_analyzed, viral_content_found, scraping_completed, profile_analyzed, template_generated, all]
                  description: Eventos que o webhook deve receber
                secret:
                  type: string
                  description: Chave secreta para assinatura HMAC (opcional)
                max_retries:
                  type: integer
                  minimum: 0
                  maximum: 10
                  default: 3
                retry_delay:
                  type: integer
                  minimum: 1
                  maximum: 300
                  default: 5
                timeout:
                  type: integer
                  minimum: 5
                  maximum: 120
                  default: 30
                metadata:
                  type: object
                  description: Metadados adicionais
              required: [url]
      responses:
        '201':
          description: Webhook registrado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'
  /webhooks/{webhook_id}:
    get:
      tags: [Webhooks]
      summary: Buscar webhook específico
      parameters:
        - name: webhook_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detalhes do webhook
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'

    put:
      tags: [Webhooks]
      summary: Atualizar webhook
      parameters:
        - name: webhook_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  format: uri
                events:
                  type: array
                  items:
                    type: string
                active:
                  type: boolean
                secret:
                  type: string
                metadata:
                  type: object
      responses:
        '200':
          description: Webhook atualizado com sucesso

    delete:
      tags: [Webhooks]
      summary: Deletar webhook
      parameters:
        - name: webhook_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Webhook deletado com sucesso
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT para autenticação de usuários
    AdminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT com privilégios de administrador
  schemas:
...
    AnalysisResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            content_id:
              type: string
            analysis_results:
              type: object
              properties:
                sentiment_analysis:
                  $ref: '#/components/schemas/SentimentAnalysis'
                visual_analysis:
                  $ref: '#/components/schemas/VisualAnalysis'
                metrics_analysis:
                  $ref: '#/components/schemas/MetricsAnalysis'
            viral_score:
              type: integer
              minimum: 0
              maximum: 100
            recommendations:
              type: array
              items:
                $ref: '#/components/schemas/Recommendation'
        timestamp:
          type: string
          format: date-time
...
    MetricsAnalysis:
      type: object
      properties:
        engagement_rate:
          type: number
        viral_velocity:
          type: number
        audience_quality:
          type: number
        growth_prediction:
          type: object
          properties:
            next_24h:
              type: number
            next_7d:
              type: number
            confidence:
              type: number
...
    TemplateResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            template_id:
              type: string
            template_name:
              type: string
            content_type:
              type: string
            viral_score:
              type: integer
            visual_structure:
              type: object
            content_formula:
              type: object
            performance_metrics:
              type: object
            extraction_metadata:
              type: object
        timestamp:
          type: string
          format: date-time
    ProfileAnalysisResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            profile_info:
              type: object
              properties:
                username:
                  type: string
                platform:
                  type: string
                followers:
                  type: integer
                following:
                  type: integer
                posts_count:
                  type: integer
                is_verified:
                  type: boolean
                engagement_potential:
                  type: integer
            content_analysis:
              type: object
              properties:
                total_posts:
                  type: integer
                viral_posts:
                  type: integer
                viral_rate:
                  type: string
                avg_engagement_rate:
                  type: number
            patterns:
              type: object
            insights:
              type: object
            templates:
              type: array
              items:
                type: object
        timestamp:
          type: string
          format: date-time
    WebhookResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            webhook_id:
              type: string
            url:
              type: string
            events:
              type: array
              items:
                type: string
            active:
              type: boolean
            created_at:
              type: string
              format: date-time
            delivery_count:
              type: integer
            failure_count:
              type: integer
        timestamp:
          type: string
          format: date-time
...
    CreateTemplateRequest:
      type: object
      properties:
        template_name:
          type: string
        content_type:
          type: string
          enum: [carousel, reel, post, story]
        visual_structure:
          type: object
        content_formula:
          type: object
        target_platforms:
          type: array
          items:
            type: string
        search_tags:
          type: array
          items:
            type: string
      required: [template_name, content_type]

    # ERROR RESPONSES
...
    ViralTrendsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            viral_content:
              type: array
              items:
                type: object
            trends_summary:
              type: object
            insights:
              type: array
              items:
                type: object
...
    TemplatesListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            templates:
              type: array
              items:
                $ref: '#/components/schemas/TemplateResponse'
            pagination:
              type: object
              properties:
                total:
                  type: integer
                limit:
                  type: integer
                offset:
                  type: integer
                has_more:
                  type: boolean
...
    ProfilesListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            profiles:
              type: array
              items:
                $ref: '#/components/schemas/ProfileAnalysisResponse'
            summary:
              type: object
...
    WebhooksListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            webhooks:
              type: array
              items:
                $ref: '#/components/schemas/WebhookResponse'
            summary:
              type: object
              properties:
                total_webhooks:
                  type: integer
                active_webhooks:
                  type: integer
                total_deliveries:
                  type: integer
  responses:
    BadRequest:
      description: Requisição inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Parâmetros obrigatórios não fornecidos"
            error_code: "MISSING_PARAMETERS"
            timestamp: "2025-01-27T10:30:00Z"
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Recurso não encontrado"
            error_code: "NOT_FOUND"
            timestamp: "2025-01-27T10:30:00Z"
...
    InternalError:
      description: Erro interno do servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Erro interno do servidor"
            error_code: "INTERNAL_ERROR"
            timestamp: "2025-01-27T10:30:00Z"
...